requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
scipy>=1.4.1
torch==2.1.0
torchvision==0.16.0
tqdm>=4.41.0
protobuf<4.21.3
ptwt==0.1.6
einops==0.7.0
# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
# Extras --------------------------------------
thop  # FLOPs computation
